+++ vortex-tensor/src/fixed_shape/metadata.rs
