+++ vortex-tensor/src/fixed_shape/mod.rs
@@ -1,0 +1,14 @@
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-FileCopyrightText: Copyright the Vortex contributors
+
+//! Fixed-shape Tensor extension type.
+
+/// The VTable for the Tensor extension type.
+#[derive(Clone, Debug, Default, PartialEq, Eq, Hash)]
+pub struct FixedShapeTensor;
+
+mod metadata;
+pub use metadata::FixedShapeTensorMetadata;
+
+mod proto;
+mod vtable;
